requirements.txt
asgiref==3.9.1
click==8.3.0
colorama==0.4.6
Django==5.2.6
django-allow-cidr==0.8.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
h11==0.16.0
inflection==0.5.1
iniconfig==2.1.0
mysqlclient==2.2.7
oracledb==3.3.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg==3.2.10
psycopg-binary==3.2.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
uvicorn==0.36.0
